RiverStead/Text-to-SQL_RAG_Chatbot/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
mysql-connector-python>=9.0.0
faiss-cpu>=1.7.0
sentence-transformers>=3.0.0
groq>=0.25.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
